requirements.txt
Django == 1.8.7
django-location-field == 1.6.1
djangorestframework == 3.3.1
Pillow == 3.0.0
selenium == 2.48.0

